feat: gnb1 헤더 api 연결
